fixes feedback
